node: Move wire code
