+++ node/src/lib.rs
@@ -4,6 +4,7 @@ pub use nakamoto_net::{Io, Link, LocalDuration, LocalTime};
 pub mod client;
 pub mod control;
 pub mod crypto;
+pub mod storage;
 
 mod address_book;
 mod address_manager;
@@ -17,7 +18,18 @@ mod logger;
 mod rad;
 mod serde_ext;
 mod service;
-mod storage;
 #[cfg(test)]
 mod test;
 mod transport;
+mod wire;
+
+pub mod prelude {
+    pub use crate::crypto::{PublicKey, Signature, Signer};
+    pub use crate::decoder::Decoder;
+    pub use crate::hash::Digest;
+    pub use crate::identity::{Did, Id};
+    pub use crate::service::filter::Filter;
+    pub use crate::service::{NodeId, Timestamp};
+    pub use crate::storage::refs::Refs;
+    pub use crate::storage::WriteStorage;
+}
