+++ node/src/service/filter.rs
@@ -1,10 +1,8 @@
-use std::io;
 use std::ops::{Deref, DerefMut};
 
-use bloomy::BloomFilter;
+pub use bloomy::BloomFilter;
 
 use crate::identity::Id;
-use crate::service::wire;
 
 /// Size in bytes of subscription bloom filter.
 pub const FILTER_SIZE: usize = 1024 * 16;
@@ -51,34 +49,8 @@ impl DerefMut for Filter {
     }
 }
 
-#[cfg(test)]
 impl From<BloomFilter<Id>> for Filter {
     fn from(bloom: BloomFilter<Id>) -> Self {
         Self(bloom)
-    }
-}
-
-impl wire::Encode for Filter {
-    fn encode<W: io::Write + ?Sized>(&self, writer: &mut W) -> Result<usize, io::Error> {
-        let mut n = 0;
-
-        n += self.0.as_bytes().encode(writer)?;
-
-        Ok(n)
-    }
-}
-
-impl wire::Decode for Filter {
-    fn decode<R: std::io::Read + ?Sized>(reader: &mut R) -> Result<Self, wire::Error> {
-        let size: wire::Size = wire::Decode::decode(reader)?;
-        if size as usize != FILTER_SIZE {
-            return Err(wire::Error::InvalidFilterSize(size as usize));
-        }
-        let bytes: [u8; FILTER_SIZE] = wire::Decode::decode(reader)?;
-        let bf = BloomFilter::from(Vec::from(bytes));
-
-        debug_assert_eq!(bf.hashes(), FILTER_HASHES);
-
-        Ok(Self(bf))
     }
 }
