+++ node/src/test/arbitrary.rs
@@ -16,13 +16,14 @@ use crate::hash;
 use crate::identity::{Delegate, Did, Doc, Id, Project};
 use crate::service::filter::{Filter, FILTER_SIZE};
 use crate::service::message::{
-    Address, Envelope, InventoryAnnouncement, Message, MessageType, NodeAnnouncement,
-    RefsAnnouncement, Subscribe,
+    Address, Envelope, InventoryAnnouncement, Message, NodeAnnouncement, RefsAnnouncement,
+    Subscribe,
 };
 use crate::service::{NodeId, Timestamp};
 use crate::storage;
 use crate::storage::refs::{Refs, SignedRefs};
 use crate::test::storage::MockStorage;
+use crate::wire::message::MessageType;
 
 use super::crypto::MockSigner;
